Menu
I've added a link for menu and created a draft of the web page. There will be twelve menu items in total. Brownies, ice cream, milkshakes and extras will all have three items each. If you could add a shopping cart, and draft out the page for the shopping cart. 

Issues: 
There seems to be a problem with the padding in the class of 'content'. The set up doesn't allow you to move the contents of the page. As a result, the 'menu' heading has stayed in place.