CSS for Menu
I have added social media icons facebook, twitter and instagram to the side of the web page. I have changed the color of the text in the header where the background was brown to make it stand out more. 

Issues: 
The social media icons are not changing in size. Will fix ASAP.