update brownie project.
I changed the sizing of the social-icons . I put the script of the thumbnail buttons back because of the mistake of taking it out.  I took the sci class out of the html code to make the icons bigger.  I also put a click me text next to the thumbnail buttons as an incentive to click on the buttons next to it.  around to thumbnail buttons I put a border- radius so it would look a little bit more neat.