commit for the menu
I worked on the menu of the project. let me know if you like it. I got stuck with changing the color of the circle. when I save my progress in atom it does not seem to change the webpage for the thing I have changed. the circle needs a different color and the section need to be a little smaller so it does not take that much space on the site.